chore: nuevo esqueleto
