+++ README.md
@@ -1,79 +1,115 @@
----
-## **NEOAZTLAN: El Duelo Cósmico**
+# NeoAztlán – Flutter Monorepo Skeleton
 
-Forja tu leyenda en NEOAZTLAN, un TCG donde la mitología ancestral se encuentra con la guerra ciberpunk.
+Este esqueleto está listo para subirse a **tres ramas** del mismo repositorio:
+- `team/flutter` – UI, navegación, widgets, estados.
+- `team/flame` – Gameplay con Flame, sistemas ECS, assets del juego.
+- `team/firebase` – Integraciones (auth, firestore, storage, functions), reglas y config.
 
-Este documento es el informe integral y la hoja de ruta para el desarrollo del proyecto "NEOAZTLAN".
-
-### **📜 Resumen Ejecutivo**
-
-"NEOAZTLAN" ha evolucionado de una iniciativa de práctica interna a una propuesta de negocio completa para un juego de cartas coleccionables (TCG) para móviles. El concepto se apalanca en la identidad de nuestra empresa para lanzar un producto con una temática única: la fusión de la mitología mesoamericana con una estética ciberpunk.
-
-Actualmente, el proyecto ha completado su fase conceptual y se encuentra en la etapa de **preproducción**, con el objetivo de desarrollar un **Prototipo Jugable Mínimo (MVP)** en los próximos tres meses.
+> La **misma base** se sube a cada rama. Después **cada equipo genera sus plataformas** localmente con `flutter create` y agrega sus dependencias. Así evitamos conflictos cruzados en nativos.
 
 ---
 
-### **🗺️ Cronograma de Desarrollo (Hoja de Ruta a 9 Meses)**
-
-El proyecto se ejecutará en dos etapas principales para asegurar un progreso medible y controlado.
-
-#### **Etapa 1: Desarrollo del MVP (Primeros 3 Meses)**
-
-El objetivo es crear una versión funcional para un solo jugador (vs. IA) que demuestre que el concepto es divertido.
-
-| Mes | Meta del Mes | Hito Clave |
-| :--- | :--- | :--- |
-| **Mes 1** | Fundación y Prototipo Básico | Prototipo Pre-Alpha con lógica de turnos y Tonal. |
-| **Mes 2** | Implementación de Mecánicas | Alpha Interna con una partida completa jugable. |
-| **Mes 3** | Pulido y UI | **MVP v1.0 Completo** (versión .apk lista). |
-
-#### **Etapa 2: Preparación para el Lanzamiento (Próximos 6 Meses)**
-
-El objetivo es convertir el MVP en una aplicación pulida, con multijugador y lista para un lanzamiento "soft".
-
-| Fase | Meta de la Fase | Hito Clave |
-| :--- | :--- | :--- |
-| **Meses 4-6** | Multijugador y Expansión | Versión Beta con PvP funcional y una nueva facción. |
-| **Meses 7-8** | Pulido y Monetización | Versión "Feature Complete" con tienda y Pase de Batalla. |
-| **Mes 9** | Lanzamiento "Soft" y Legal | ¡Aplicación Jugable y Lanzable en un mercado de prueba! |
+## 1) Requisitos
+- Flutter 3.22+ (`flutter --version`)
+- Dart SDK incluido en Flutter
+- Git
+- (opcional) Android Studio / Xcode según plataforma
 
 ---
 
-### **👥 Perfiles del Equipo de Desarrollo (El "Dream Team")**
+## 2) Crear plataformas por rama (local)
+> Las carpetas nativas `android/ ios/ windows/ macos/ linux/ web/` **no se versionan** en este esqueleto inicial. Cada equipo las genera en su rama.
 
-Para ejecutar este cronograma, se requiere un equipo central ágil con los siguientes perfiles clave:
+```bash
+# Ejemplo para generar TODAS las plataformas
+flutter create . --platforms=android,ios,windows,macos,linux,web
+```
 
-| Rol | Título Temático | Responsabilidades Principales |
-| :--- | :--- | :--- |
-| **Líder de Proyecto** | El "Tlatoani" | Supervisar la visión, gestionar el cronograma y comunicar los avances. |
-| **Desarrollador de UI** | El "Códice Viviente" | Construir la interfaz de la app (menús, tienda, etc.) en Flutter. |
-| **Desarrollador de Juego** | El "Ingeniero de Tonal" | Programar la lógica del tablero de juego y el combate en Flame. |
-| **Diseñador UI/UX y Artista** | El "Tlacuilo" | Definir la dirección de arte, diseñar las interfaces en Figma y crear el arte. |
+- **Flutter team**: genera todas o las de escritorio/web según necesiten para UI.
+- **Flame team**: genera las plataformas objetivo del juego.
+- **Firebase team**: puede enfocarse en Android/iOS/Web.
 
-**Nota:** Roles adicionales como Legal y Sonido se consideran servicios externos a contratar en la Etapa 2.
-
----
-
-### **💰 Análisis de Inversión (Capital Directo)**
-
-A continuación, se presenta un desglose de la inversión de capital necesaria, excluyendo los costos de oportunidad del equipo interno.
-
-#### **Inversión para la Etapa 1: MVP (Primeros 3 Meses)**
-
-- **Objetivo:** Validar el concepto con una inversión de capital mínima.
-- **Capital Requerido Total:** ~$1,700 MXN (Principalmente para cuentas de desarrollador y herramientas opcionales).
-
-#### **Inversión para la Etapa 2: Preparación para el Lanzamiento (Próximos 6 Meses)**
-
-- **Objetivo:** Transformar el MVP en un producto comercialmente viable (esta inversión se realiza únicamente tras el éxito del MVP).
-- **Capital Requerido Total:** $96,000 - $186,000 MXN (Para assets profesionales, marketing inicial y registro de marca).
+> Si una rama necesita subir los nativos (recomendado en etapas de build), simplemente elimina esas rutas del `.gitignore` y commitea.
 
 ---
 
-### **⚖️ Licencia**
+## 3) Dependencias por rama (sugerencias)
 
-Este proyecto y todo su contenido están protegidos bajo una licencia de derechos de autor.
+### team/flutter
+```bash
+flutter pub add go_router riverpod flutter_hooks
+```
 
-**Copyright © 2024-2025 Teteocan Technologies. Todos los derechos reservados.**
+### team/flame
+```bash
+flutter pub add flame flame_audio flame_tiled
+```
 
-Para más detalles, por favor, consulta el archivo `LICENSE.md` en este repositorio.
+### team/firebase
+```bash
+flutter pub add firebase_core firebase_auth cloud_firestore firebase_storage
+# Web setup: flutterfire configure (si lo requieren)
+```
+
+---
+
+## 4) Scripts útiles
+
+- Generar todas las plataformas:
+  ```bash
+  bash scripts/setup_all_platforms.sh
+  ```
+
+- Limpiar y reconstruir:
+  ```bash
+  flutter clean && flutter pub get
+  ```
+
+---
+
+## 5) Estructura
+
+```
+.
+├── lib/
+│  └── main.dart
+├── test/
+│  └── widget_test.dart
+├── scripts/
+│  └── setup_all_platforms.sh
+├── analysis_options.yaml
+├── pubspec.yaml
+├── .gitignore
+└── README.md
+```
+
+---
+
+## 6) Convenciones de ramas
+
+- `main`: solo releases estables.
+- `dev`: integración temporal si lo necesitan.
+- `team/flutter`, `team/flame`, `team/firebase`: trabajo diario de cada equipo.
+
+Para **PRs entre equipos**, abrir contra `dev` y luego `dev -> main` cuando haya hitos.
+
+---
+
+## 7) Checklist por equipo
+
+**Flutter**
+- [ ] Generar plataformas necesarias
+- [ ] Definir rutas (go_router)
+- [ ] Layout base y tema
+
+**Flame**
+- [ ] Generar plataformas de target
+- [ ] Crear `Game` base en `lib/game/`
+- [ ] Cargar assets y sistemas
+
+**Firebase**
+- [ ] `flutterfire configure` si aplicable
+- [ ] Inicializar Firebase en `main.dart` condicional por flavor
+- [ ] Reglas y emuladores
+
+¡Listo para trabajar! 🚀
