--- flame/README.md
@@ -1,2 +1,0 @@
-# flame_module
-Módulo local para el juego con Flame. La app Flutter lo consume por path dependency.
