--- flame/lib/neo_game.dart
@@ -1,8 +1,0 @@
-import 'package:flame/game.dart';
-
-class NeoGame extends FlameGame {
-  @override
-  Future<void> onLoad() async {
-    // TODO: carga sprites, audio, escenas...
-  }
-}
