--- flame/pubspec.yaml
@@ -1,13 +1,0 @@
-name: flame_module
-description: Local Flame module for NeoAztlan
-version: 0.0.1
-environment:
-  sdk: '>=3.4.0 <4.0.0'
-
-dependencies:
-  flame: ^1.15.0
-  flutter:
-    sdk: flutter
-
-dev_dependencies:
-  flutter_lints: ^4.0.0
