--- flutter/analysis_options.yaml
@@ -1,6 +1,0 @@
-include: package:flutter_lints/flutter.yaml
-linter:
-  rules:
-    avoid_print: true
-    prefer_const_constructors: true
-    prefer_final_locals: true
