--- flutter/lib/core/theme.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-ThemeData buildTheme() {
-  return ThemeData(
-    colorScheme: ColorScheme.fromSeed(seedColor: Color(0xFF10593E)),
-    useMaterial3: true,
-  );
-}
