--- flutter/lib/firebase_options.dart
@@ -1,11 +1,0 @@
-// Archivo placeholder.
-// Genera el real ejecutando: 
-//   dart pub global activate flutterfire_cli
-//   flutterfire configure --platforms=android,ios,web
-//
-// NO subas credenciales aquí. Este archivo es auto-generado.
-
-class DefaultFirebaseOptions {
-  static dynamic get currentPlatform =>
-      throw UnimplementedError('Ejecuta flutterfire configure para generar firebase_options.dart');
-}
