+++ pubspec.yaml
@@ -1,0 +1,21 @@
+name: neoaztlan
+description: Base Flutter app skeleton for the multi-team project (Flutter/Flame/Firebase).
+publish_to: "none"
+
+version: 0.1.0+1
+
+environment:
+  sdk: ">=3.4.0 <4.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  cupertino_icons: ^1.0.8
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^4.0.0
+
+flutter:
+  uses-material-design: true
